src/services/S3Service/__tests__/getClient.test.ts
import getClient from '../getClient';
import logger from '../../../utils/logger';

jest.mock('../../../utils/logger');

describe('getClient', () => {
  afterEach(() => {
    jest.clearAllMocks();
  });

  it('should create a new client if singleton connection is not yet created', () => {
    const region = 'us-west-2';
    const singletonConn = 'default';

    const client = getClient({ region, singletonConn });

    expect(logger.debug).toHaveBeenCalledTimes(1);
    expect(logger.debug).toHaveBeenCalledWith(
      'lesgo.services.S3Service.getClient::NEW_CLIENT',
      {
        client,
        region: 'us-west-2',
      }
    );
  });

  it('should return the existing client if singleton connection is already created', () => {
    const region = 'us-west-2';
    const singletonConn = 'default';

    const client1 = getClient({ region, singletonConn });
    getClient({ region, singletonConn });

    expect(logger.debug).toHaveBeenCalledTimes(2);
    expect(logger.debug).toHaveBeenCalledWith(
      'lesgo.services.S3Service.getClient::REUSE_CLIENT_SINGLETON',
      {
        client: client1,
        region: 'us-west-2',
      }
    );
  });
});
